+++ packages/graph-explorer/src/connector/gremlin/mappers/mapApiVertex.ts
@@ -4,8 +4,11 @@ import parsePropertiesValues from "./parsePropertiesValues";
 import { extractRawId } from "./extractRawId";
 
 const mapApiVertex = (apiVertex: GVertex): Vertex => {
-  const labels = apiVertex["@value"].label.split("::");
-  const vt = labels[0];
+  // Split multi-label from Neptune and filter out empty strings
+  const labels = apiVertex["@value"].label.split("::").filter(Boolean);
+
+  // Check for empty labels, which is possible with some databases
+  const vt = labels[0] ?? "";
   const isFragment = apiVertex["@value"].properties == null;
 
   return {